+++ notification/src/epilot/sdk.py
@@ -5,7 +5,7 @@ from .notification import Notification
 from .sdkconfiguration import SDKConfiguration
 from epilot import utils
 from epilot.models import shared
-from typing import Callable, Dict, Union
+from typing import Dict
 
 class Epilot:
     r"""Notification API: Notification API for epilot 360"""
@@ -15,7 +15,7 @@ class Epilot:
     sdk_configuration: SDKConfiguration
 
     def __init__(self,
-                 epilot_auth: Union[str,Callable[[], str]],
+                 epilot_auth: str ,
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: Dict[str, str] = None,